% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_data_url.R
\name{construct_data_url}
\alias{construct_data_url}
\title{Construct URL request for data}
\usage{
construct_data_url(
  station_id = "cht01",
  time_step = "hourly",
  date_from = lubridate::today() - 5,
  date_to = "now",
  network = "coagmet"
)
}
\arguments{
\item{station_id}{ID of weather station; default 'cht01'. Use 'all' to get data for all stations}

\item{time_step}{Frequency of data (5min, hourly, daily, latest); default 'hourly'}

\item{date_from}{Start date of data to get in yyyy-mm-dd format: default is 5 days ago}

\item{date_to}{End date of data in yyyy-mm-dd format; defaut 'now'}

\item{network}{Network :either "coagmet" or "nw" (northern water). Default is coagmet}
}
\value{
data_url : UR to request data from CoAgMet API
}
\description{
Construct URL request for data
}
\examples{
url <- construct_data_url()

}
