% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_coagmet_data_csv.R
\name{fetch_coagmet_data_csv}
\alias{fetch_coagmet_data_csv}
\title{Download CoAgMet data (csv) for specified url}
\usage{
fetch_coagmet_data_csv(data_url)
}
\arguments{
\item{data_url}{API url for data request}
}
\value{
Data frame of requested CoAgMet data
}
\description{
Download CoAgMet data (csv) for specified url
}
\seealso{
\code{\link[=construct_data_url]{construct_data_url()}}
}
