% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coagmet_data.R
\name{get_coagmet_data}
\alias{get_coagmet_data}
\title{Get and process CoAgMet weather station (csv) data}
\usage{
get_coagmet_data(
  station_id = "cht01",
  time_step = c("hourly", "5min", "daily", "latest"),
  date_from = lubridate::today() - 5,
  date_to = "now",
  network = c("coagmet", "nw")
)
}
\arguments{
\item{station_id}{ID of weather station. Default = 'cht01'. Use "all" to retrieve data for all stations}

\item{time_step}{Frequency of data (5min, hourly, daily, latest); default 'hourly'}

\item{date_from}{Start date of data to get in yyyy-mm-dd format: default is 5 days ago}

\item{date_to}{End date of data in yyyy-mm-dd format; defaut = 'now'}

\item{network}{Network :either "coagmet" or "nw" (northern water). Default is coagmet}
}
\value{
Data frame of processed CoAgMet data
}
\description{
Get and process CoAgMet weather station (csv) data
}
\examples{
df <- get_coagmet_data(station_id = "cht01", time_step = "daily")
}
