% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coagmet_data_one_station.R
\name{get_coagmet_data_one_station}
\alias{get_coagmet_data_one_station}
\title{Get and process CoAgMet data for one station}
\usage{
get_coagmet_data_one_station(
  station_id = "cht01",
  time_step = "hourly",
  date_from = lubridate::today() - 5,
  date_to = "now"
)
}
\arguments{
\item{station_id}{ID of weather station; default 'cht01'}

\item{time_step}{Frequency of data (5min, hourly, daily); default 'hourly'}

\item{date_from}{Start date of data to get in yyyy-mm-dd format: default is 5 days ago}

\item{date_to}{End date of data in yyyy-mm-dd format; defaut 'now'}
}
\value{
Data frame of processed CoAgMet data for one station
}
\description{
Get and process CoAgMet data for one station
}
\examples{
df <- get_coagmet_data_one_station(station_id = "cht01", time_step = "daily")
}
