% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coagmet_plotly.R
\name{plot_coagmet_plotly}
\alias{plot_coagmet_plotly}
\title{Make plotly plot of one variable}
\usage{
plot_coagmet_plotly(df, var_to_plot = "air_temp")
}
\arguments{
\item{df}{Dataframe of processed CoAgMet data}

\item{var_to_plot}{Variable to plot; default is air_temp}
}
\value{
g: Plotly plot object
}
\description{
Make plotly plot of one variable
}
