% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coagmet_data_csv_daily.R
\name{process_coagmet_data_csv_daily}
\alias{process_coagmet_data_csv_daily}
\title{Process daily CoAgMet (csv) data returned from API}
\usage{
process_coagmet_data_csv_daily(df)
}
\arguments{
\item{df}{Data frame of CoAgMet data as returned by fetch_coagmet_data_csv.R}
}
\value{
Processed data frame of daily CoAgMet data
}
\description{
Process daily CoAgMet (csv) data returned from API
}
\seealso{
\code{\link[=fetch_coagmet_data_csv]{fetch_coagmet_data_csv()}}
}
