% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcoagmet-package.R
\docType{package}
\name{rcoagmet-package}
\alias{rcoagmet}
\alias{rcoagmet-package}
\title{rcoagmet: API Wrapper for CoAgMet Weather Stations}
\description{
API Wrapper for CoAgMet Weather Stations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andypicke/rcoagmet}
  \item Report bugs at \url{https://github.com/andypicke/rcoagmet/issues}
}

}
\author{
\strong{Maintainer}: Andy Pickering \email{andypicke@gmail.com} (\href{https://orcid.org/0000-0002-5724-7042}{ORCID})

}
\keyword{internal}
