% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_coagmet_data_csv.R
\name{fetch_coagmet_data_csv}
\alias{fetch_coagmet_data_csv}
\title{fetch_coagmet_data_csv}
\usage{
fetch_coagmet_data_csv(data_url)
}
\arguments{
\item{data_url}{API url for data request}
}
\value{
Data frame of CoAgMet data
}
\description{
Fetch csv data from CoAgMet API for specified url
}
\author{
Andy Pickering
}
