% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_coagmet_station.R
\name{find_closest_coagmet_station}
\alias{find_closest_coagmet_station}
\title{Find closest CoAgMet station to given location (using Haversine formula)}
\usage{
find_closest_coagmet_station(xlat, xlon, network = c("coagmet", "nw"))
}
\arguments{
\item{xlat}{latitidue (degN)}

\item{xlon}{longitude (degE)}

\item{network}{Network :either "coagmet" or "nw" (northern water). Default is coagmet}
}
\value{
closest_station: 1 row Dataframe of info for closest station, including distance in meters
}
\description{
Find closest CoAgMet station to given location (using Haversine formula)
}
\examples{
closest <- find_closest_coagmet_station(xlat = 39.74, xlon = -104.99)

}
