% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coagmet_meta.R
\name{get_coagmet_meta}
\alias{get_coagmet_meta}
\title{Get Metadata for CoAgMet Weather Stations}
\usage{
get_coagmet_meta(station_id = "all")
}
\arguments{
\item{station_id}{ID of station to retrieve info for. Default is all stations}
}
\value{
Data frame of metadata for all CoAgMet Weather Stations
}
\description{
Get Metadata for CoAgMet Weather Stations
}
\author{
Andy Pickering
}
