% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coagmet_data_csv.R
\name{process_coagmet_data_csv}
\alias{process_coagmet_data_csv}
\title{Process CoAgMet data returned from API}
\usage{
process_coagmet_data_csv(df)
}
\arguments{
\item{df}{Data frame of CoAgMet data as returned by fetch_coagmet_data_csv.R}
}
\value{
Processed data frame of CoAgMet data
}
\description{
Process CoAgMet data returned from API
}
